+++ scripts/dzVents/runtime/integration-tests/stage1.lua
@@ -5,6 +5,10 @@ local err = function(msg)
 	log(msg, dz.LOG_ERROR)
 end
 
+local tstMsg = function(msg, res)
+	print('Stage: 1, ' .. msg .. ': ' .. tostring(res and 'OK' or 'FAILED'))
+end
+
 local expectEql = function (attr, test, marker)
 	if (attr ~= test) then
 		local msg = tostring(attr) .. '~=' .. tostring(test)
@@ -45,6 +49,7 @@ local testAirQuality = function(name)
 		["deviceType"] = "Air Quality";
 	})
 	dev.updateAirQuality(1234)
+	tstMsg('Test air quality device', res)
 	return res
 end
 
@@ -68,18 +73,21 @@ local testSwitch = function(name)
 		["description"] = 'desc vdSwitch'
 	})
 	dev.switchOn().afterSec(1)
+	tstMsg('Test switch device', res)
 	return res
 end
 
 local testDimmer = function(name)
 	local dev = dz.devices(name)
+
 	local res = true
 	res = res and checkAttributes(dev, {
 		["id"] = 39,
 		["name"] = name,
 		["maxDimLevel"] = 100,
-		["lastLevel"] = 0,
-		["state"] = "Off",
+		['level'] = 33,
+		["lastLevel"] = 33, -- this script is NOT triggered by the dimmer so lastLevel is the current level
+		["state"] = "On",
 		["deviceSubType"] = "Switch";
 		["hardwareType"] = "Dummy (Does nothing, use for virtual switches only)";
 		["hardwareName"] = "dummy";
@@ -93,7 +101,9 @@ local testDimmer = function(name)
 		["deviceType"] = "Light/Switch";
 		["description"] = "desc vdSwitchDimmer"
 	})
+
 	dev.dimTo(75).afterSec(1)
+	tstMsg('Test dimmer', res)
 	return res
 end
 
@@ -116,6 +126,7 @@ local testAlert = function(name)
 		["deviceType"] = "General";
 	})
 	dev.updateAlertSensor(dz.ALERTLEVEL_RED, 'Hey I am red')
+	tstMsg('Test alert sensor device', res)
 	return res
 end
 
@@ -140,7 +151,7 @@ local testAmpere3 = function(name)
 	})
 
 	dev.updateCurrent(123, 456, 789)
-
+	tstMsg('Test ampere 3 device', res)
 	return res
 end
 
@@ -166,6 +177,7 @@ local testAmpere1 = function(name)
 	})
 
 	dev.updateCurrent(123)
+	tstMsg('Test ampere 1 device', res)
 
 	return res
 end
@@ -189,6 +201,7 @@ local testBarometer = function(name)
 		["deviceType"] = "General";
 	})
 	dev.updateBarometer(1234, dz.BARO_THUNDERSTORM)
+	tstMsg('Test barometer device', res)
 	return res
 end
 
@@ -213,6 +226,7 @@ local testCounter = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateCounter(1234)
+	tstMsg('Test counter device', res)
 	return res
 end
 
@@ -237,6 +251,7 @@ local testCounterIncremental = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateCounter(1234)
+	tstMsg('Test counter incremental device', res)
 	return res
 end
 
@@ -260,6 +275,7 @@ local testCustomSensor = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateCustomSensor(1234)
+	tstMsg('Test custom sensor device', res)
 	return res
 end
 
@@ -282,6 +298,7 @@ local testDistance = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateDistance(42.44)
+	tstMsg('Test distance device', res)
 	return res
 end
 
@@ -306,6 +323,7 @@ local testElectricInstanceCounter = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateElectricity(10, 20)
+	tstMsg('Test electric instance counter device', res)
 	return res
 end
 
@@ -328,6 +346,7 @@ local testGas = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateGas(6000) -- this won't have a direct effect
+	tstMsg('Test gas device', res)
 	return res
 end
 
@@ -350,6 +369,7 @@ local testHumidity = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateHumidity(88, dz.HUM_WET)
+	tstMsg('Test humidity device', res)
 	return res
 end
 
@@ -371,6 +391,7 @@ local testLeafWetness = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateWetness(55)
+	tstMsg('Test leaf wetness device', res)
 	return res
 end
 
@@ -392,6 +413,7 @@ local testLux = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateLux(355)
+	tstMsg('Test lux device', res)
 	return res
 end
 
@@ -421,6 +443,7 @@ local testP1SmartMeter = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateP1(10, 20, 100, 200, 666, 777)
+	tstMsg('Test p1 smart meter device', res)
 	return res
 end
 
@@ -442,6 +465,7 @@ local testPercentage = function(name)
 		["timedOut"] = false;
 	})
 	dev.updatePercentage(99.99)
+	tstMsg('Test percentage device', res)
 	return res
 end
 
@@ -463,6 +487,7 @@ local testPressureBar = function(name)
 		["timedOut"] = false;
 	})
 	dev.updatePressure(88)
+	tstMsg('Test pressure device', res)
 	return res
 end
 
@@ -485,6 +510,7 @@ local testRain = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateRain(3000, 6660)
+	tstMsg('Test rain device', res)
 	return res
 end
 
@@ -507,6 +533,7 @@ local testRGB = function(name)
 		["timedOut"] = false;
 	})
 	dev.dimTo(15)
+	tstMsg('Test rgb device', res)
 	return res
 end
 
@@ -529,6 +556,7 @@ local testRGBW = function(name)
 		["timedOut"] = false;
 	})
 	dev.dimTo(15)
+	tstMsg('Test rgbw device', res)
 	return res
 end
 
@@ -550,6 +578,7 @@ local testScaleWeight = function(name)
 		["timedOut"] = false;
 	})
 	dev.updateWeight(33.5)
+	tstMsg('Test scale weight device', res)
 	return res
 end
 
@@ -579,6 +608,7 @@ local testSelectorSwitch = function(name)
 	res = res and expectEql('Level3', dev.levelNames[4])
 
 	dev.switchSelector(30) -- level3
+	tstMsg('Test selector switch device', res)
 	return res
 end
 
@@ -601,6 +631,7 @@ local testSoilMoisture = function(name)
 	})
 
 	dev.updateSoilMoisture(34)
+	tstMsg('Test soil moisture device', res)
 	return res
 end
 
@@ -623,6 +654,7 @@ local testSolarRadiation = function(name)
 	})
 
 	dev.updateRadiation(34)
+	tstMsg('Test solar radiation device', res)
 	return res
 end
 
@@ -645,6 +677,7 @@ local testSoundLevel = function(name)
 	})
 
 	dev.updateSoundLevel(120)
+	tstMsg('Test sound level device', res)
 	return res
 end
 
@@ -667,6 +700,7 @@ local testTemperature = function(name)
 	})
 
 	dev.updateTemperature(120)
+	tstMsg('Test temperature device', res)
 	return res
 end
 
@@ -691,6 +725,7 @@ local testTempHum = function(name)
 	})
 
 	dev.updateTempHum(34, 88, dz.HUM_WET)
+	tstMsg('Test temperature+humidity device', res)
 	return res
 end
 
@@ -718,6 +753,7 @@ local testTempHumBaro = function(name)
 	})
 
 	dev.updateTempHumBaro(34, 88, dz.HUM_WET, 1033, dz.BARO_PARTLYCLOUDY)
+	tstMsg('Test temperature+humidity+barometer device', res)
 	return res
 end
 
@@ -740,6 +776,7 @@ local testText = function(name)
 	})
 
 	dev.updateText("Oh my Darwin, what a lot of tests!")
+	tstMsg('Test text device', res)
 	return res
 end
 
@@ -762,6 +799,7 @@ local testThermostatSetpoint = function(name)
 	})
 
 	dev.updateSetPoint(22)
+	tstMsg('Test thermostat device', res)
 	return res
 end
 
@@ -784,6 +822,7 @@ local testUsageElectric = function(name)
 	})
 
 	dev.updateEnergy(1922)
+	tstMsg('Test usage electric device', res)
 	return res
 end
 
@@ -806,6 +845,7 @@ local testUV = function(name)
 	})
 
 	dev.updateUV(12.33)
+	tstMsg('Test uv device', res)
 	return res
 end
 
@@ -828,6 +868,7 @@ local testVisibility = function(name)
 	})
 
 	dev.updateVisibility(1)
+	tstMsg('Test visibility device', res)
 	return res
 end
 
@@ -850,6 +891,7 @@ local testVoltage = function(name)
 	})
 
 	dev.updateVoltage(220)
+	tstMsg('Test voltage device', res)
 	return res
 end
 
@@ -872,6 +914,7 @@ local testWaterflow = function(name)
 	})
 
 	dev.updateWaterflow(15)
+	tstMsg('Test waterflow device', res)
 	return res
 end
 
@@ -899,6 +942,7 @@ local testWind = function(name, id, subType)
 	})
 
 	dev.updateWind(120, 'SW', 55, 66, 77, 88)
+	tstMsg('Test wind device', res)
 	return res
 end
 
@@ -908,25 +952,28 @@ local testScene = function(name)
 	res = res and checkAttributes(scene, {
 		["id"] = 1,
 		["name"] = name,
-		['state'] = 'On',
+		['state'] = 'Off',
 		['baseType'] = 'scene'
 	})
 
 	scene.switchOn()
+	tstMsg('Test scene', res)
 	return res
 end
 
 local testGroup = function(name)
 	local group = dz.groups(name)
 	local res = true
+
 	res = res and checkAttributes(group, {
 		["id"] = 2,
 		["name"] = name,
-		['state'] = 'On',
+		['state'] = 'Off',
 		['baseType'] = 'group'
 	})
 
-	group.switchOff()
+	group.switchOn()
+	tstMsg('Test group', res)
 	return res
 end
 
@@ -940,6 +987,7 @@ local testVariableInt = function(name)
 	})
 
 	var.set(43)
+	tstMsg('Test variable: int', res)
 	return res
 end
 
@@ -953,6 +1001,7 @@ local testVariableFloat = function(name)
 	})
 
 	var.set(43)
+	tstMsg('Test variable: float', res)
 	return res
 end
 
@@ -966,6 +1015,7 @@ local testVariableString = function(name)
 	})
 
 	var.set('Zork is a dork')
+	tstMsg('Test variable: string', res)
 	return res
 end
 
@@ -981,6 +1031,7 @@ local testVariableDate = function(name)
 	res = res and expectEql(var.date.day, 31)
 	res = res and expectEql(var.date.year, 2017)
 	var.set('20/11/2016');
+	tstMsg('Test variable: date', res)
 	return res
 end
 
@@ -995,6 +1046,44 @@ local testVariableTime = function(name)
 	res = res and expectEql(var.time.hour, 23)
 	res = res and expectEql(var.time.min, 59)
 	var.set('09:54');
+	tstMsg('Test variable: time', res)
+	return res
+end
+
+local storeLastUpdates = function()
+
+	dz.globalData.stage1Time = dz.time.raw
+
+end
+
+local testLastUpdates = function()
+
+	local Time = require('Time')
+	local results = true
+	local now = dz.time.secondsSinceMidnight
+
+	results = dz.devices().reduce(function(acc, device)
+
+		if (device.name ~= 'endResult' and device.name ~= 'stage1Trigger' and device.name ~= 'stage2Trigger') then
+
+			local delta = now - device.lastUpdate.secondsSinceMidnight
+
+			-- test if lastUpdate for the device is close to domoticz time
+			acc = acc and (delta <= 1)
+			expectEql(true, acc, device.name .. ' lastUpdate is not correctly set')
+		end
+
+		return acc
+	end, results)
+
+	tstMsg('Test lastUpdates', results)
+	return results
+end
+
+local testSecurity = function()
+	local res = true
+	res = res and expectEql(dz.security, dz.SECURITY_DISARMED)
+	dz.devices('secPanel').armAway()
 	return res
 end
 
@@ -1049,13 +1138,17 @@ return {
 		res = res and testWaterflow("vdWaterflow")
 		res = res and testWind('vdWind', 37, "WTGR800")
 		res = res and testWind('vdWindTempChill', 38, "TFA")
---		res = res and testScene('scScene')
---		res = res and testGroup('gpGroup')
+		res = res and testScene('scScene')
+		res = res and testGroup('gpGroup')
 		res = res and testVariableInt('varInteger')
 		res = res and testVariableFloat('varFloat')
 		res = res and testVariableString('varString')
 		res = res and testVariableDate('varDate')
 		res = res and testVariableTime('varTime')
+		res = res and testLastUpdates()
+		res = res and testSecurity()
+
+		storeLastUpdates()
 
 		log('Finishing stage 1')
 		if (not res) then
@@ -1064,6 +1157,6 @@ return {
 			log('Results stage 1: SUCCEEDED')
 		end
 
-		dz.devices('stage2Trigger').switchOn().afterSec(2)
+		dz.devices('stage2Trigger').switchOn().afterSec(4)
 	end
 }