+++ scripts/dzVents/runtime/tests/tstData.lua
@@ -23,9 +23,10 @@ local testData = {
 				["hardwareType"] = "ht1",
 				["hardwareTypeValue"] = 1;
 				["hardwareID"] = 1,
+				['unit'] = 1,
 				heatingMode = '3'
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__1",
 			["rawData"] = {
 				[1] = "1",
 				[2] = '2',
@@ -53,12 +54,13 @@ local testData = {
 				["hardwareType"] = "ht2",
 				["hardwareTypeValue"] = 2;
 				["hardwareID"] = 2,
+				['unit'] = 1,
 				temperature = 12,
 				dewpoint = 66,
 				humidity = 67,
 				--lux = 4
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__2",
 			["rawData"] = {
 				[1] = "4",
 				[2] = '5',
@@ -86,10 +88,11 @@ local testData = {
 				["hardwareType"] = "ht3",
 				["hardwareTypeValue"] = 3;
 				["hardwareID"] = 3,
+				['unit'] = 1,
 				WActual = 11,
 				uv = 23
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__3",
 			["rawData"] = {
 				[1] = "7",
 				[2] = '8',
@@ -121,13 +124,14 @@ local testData = {
 				["hardwareType"] = "ht4",
 				["hardwareTypeValue"] = 4;
 				["hardwareID"] = 4,
+				['unit'] = 1,
 				utility = 123,
 				weather = 'Nice',
 				rainLastHour = 12,
 				rain = 666,
 				setPoint = 10
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__4",
 			["rawData"] = {
 				[1] = "10",
 				[2] = '11',
@@ -157,13 +161,14 @@ local testData = {
 				["hardwareType"] = "ht5",
 				["hardwareTypeValue"] = 5;
 				["hardwareID"] = 0,
+				['unit'] = 1,
 				WhTotal = 14,
 				WActual = 13,
 				WhToday = 1.234,
 				level = 10,
 				counterTotal = '567 kWh'
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__5",
 			["rawData"] = {
 				[1] = "13",
 				[2] = '14',
@@ -190,9 +195,10 @@ local testData = {
 				["hardwareType"] = "ht4",
 				["hardwareTypeValue"] = 4;
 				["hardwareID"] = 4,
+				['unit'] = 1,
 				["value"] = 16.5, -- ?
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__6",
 			["rawData"] = {},
 			["baseType"] = "device";
 			["changed"] = true;
@@ -216,9 +222,10 @@ local testData = {
 				["hardwareTypeValue"] = 0;
 				["hardwareID"] = 0,
 				["_state"] = 16.5,
+				['unit'] = 1,
 				WActual = 16.5
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__7",
 			["rawData"] = {
 				[1] = "16.5"
 			},
@@ -244,9 +251,10 @@ local testData = {
 				["hardwareTypeValue"] = 0;
 				["hardwareID"] = 0,
 				["_state"] = 16.5,
+				['unit'] = 1,
 				text = 'Blah'
 			};
-			["deviceID"] = "",
+			["deviceID"] = "__8",
 			["rawData"] = {
 				[1] = "16.5"
 			},
@@ -255,8 +263,69 @@ local testData = {
 			["changedAttribute"] = nil --tbd
 		},
 
-		--- vars
 		[9] = {
+			["id"] = 9,
+			["name"] = "device9", -- double name
+			["description"] = "";
+			["batteryLevel"] = 10,
+			["signalLevel"] = 10,
+			["subType"] = "Text";
+			["deviceType"] = "General",
+			["timedOut"] = true,
+			["switchType"] = "",
+			["switchTypeValue"] = 0,
+			["lastUpdate"] = "2017-04-18 20:15:23";
+			["data"] = {
+				["hardwareName"] = "",
+				["hardwareType"] = "",
+				["hardwareTypeValue"] = 0;
+				["hardwareID"] = 0,
+				["_state"] = 16.5,
+				['unit'] = 1,
+				text = 'Blah'
+			};
+			["deviceID"] = "__9",
+			["rawData"] = {
+				[1] = "16.5"
+			},
+			["baseType"] = "device";
+			["changed"] = true;
+			["changedAttribute"] = nil --tbd
+		},
+
+		[10] = {
+			["id"] = 10,
+			["name"] = "device9", -- double name
+			["description"] = "";
+			["batteryLevel"] = 10,
+			["signalLevel"] = 10,
+			["subType"] = "Text";
+			["deviceType"] = "General",
+			["timedOut"] = true,
+			["switchType"] = "",
+			["switchTypeValue"] = 0,
+			["lastUpdate"] = "2017-04-18 20:15:23";
+			["data"] = {
+				["hardwareName"] = "",
+				["hardwareType"] = "",
+				["hardwareTypeValue"] = 0;
+				["hardwareID"] = 0,
+				["_state"] = 16.5,
+				['unit'] = 1,
+				text = 'Blah'
+			};
+			["deviceID"] = "__10",
+			["rawData"] = {
+				[1] = "16.5"
+			},
+			["baseType"] = "device";
+			["changed"] = true;
+			["changedAttribute"] = nil --tbd
+		},
+
+
+		--- vars
+		[11] = {
 			["id"] = 1,
 			["name"] = "x",
 			["changed"] = true,
@@ -267,7 +336,7 @@ local testData = {
 				["value"] = 1
 			}
 		},
-		[10] = {
+		[12] = {
 			["id"] = 2,
 			["name"] = "y",
 			["changed"] = false,
@@ -276,7 +345,7 @@ local testData = {
 			["lastUpdate"] = "2017-04-18 20:16:23";
 			data = { ["value"] = 2.3 }
 		},
-		[11] = {
+		[13] = {
 			["id"] = 3,
 			["name"] = "z",
 			["changed"] = true,
@@ -285,7 +354,7 @@ local testData = {
 			["lastUpdate"] = "2017-04-18 20:16:23";
 			data = { ["value"] = 'some value'}
 		},
-		[12] = {
+		[14] = {
 			["id"] = 4,
 			["name"] = "a",
 			["changed"] = true,
@@ -294,7 +363,7 @@ local testData = {
 			["lastUpdate"] = "2017-04-18 20:16:23";
 			data = { ["value"] = '3/12/2017' }
 		},
-		[13] = {
+		[15] = {
 			["id"] = 5,
 			["name"] = "b",
 			["changed"] = true,
@@ -303,9 +372,17 @@ local testData = {
 			["lastUpdate"] = "2017-04-18 20:16:23";
 			data = { ["value"] = '19:34' }
 		},
-
+		[16] = {
+			["id"] = 6,
+			["name"] = "var with spaces",
+			["changed"] = true,
+			["variableType"] = 'string',
+			["baseType"] = "uservariable";
+			["lastUpdate"] = "2017-04-18 20:16:23";
+			data = { ["value"] = 'some value' }
+		},
 		--- scenes and groups
-		[14] = {
+		[17] = {
 			["id"] = 1;
 			["baseType"] = "scene";
 			["description"] = 'Descr scene 1',
@@ -315,7 +392,7 @@ local testData = {
 			},
 			["lastUpdate"] = "2017-04-18 15:31:19";
 		},
-		[15] = {
+		[18] = {
 			["id"] = 3;
 			["baseType"] = "group";
 			["description"] = 'Descr group 1',
@@ -325,7 +402,7 @@ local testData = {
 			},
 			["lastUpdate"] = "2017-04-18 15:31:26";
 		},
-		[16] = {
+		[19] = {
 			["id"] = 2;
 			["baseType"] = "scene";
 			["description"] = 'Descr scene 2',
@@ -335,7 +412,7 @@ local testData = {
 			},
 			["lastUpdate"] = "2017-04-19 20:31:50";
 		},
-		[17] = {
+		[20] = {
 			["id"] = 4;
 			["baseType"] = "group";
 			["description"] = 'Descr group 2',